rv-trip-extractor/requirements.txt
youtube-transcript-api
google-api-python-client
openai
supabase
python-dotenv
requests
